- Refactored admin-global-search-autocomplete.php endpoint: -- Don't parse local.xml, use helper instead -- Implemented stemmers, latin transform filters -- Implemented std, prefix, suffix, brand, keyword, language analyzers

+++ app/code/community/Clean/ElasticSearch/Model/Index.php
@@ -2,6 +2,26 @@
 
 class Clean_ElasticSearch_Model_Index extends Varien_Object
 {
+    /**
+     * @var Clean_ElasticSearch_Helper_Data
+     */
+    protected $_helper;
+
+    /**
+     * @var Elastica\Index
+     */
+    protected $_index;
+
+    public function _construct()
+    {
+        $this->_helper = Mage::helper('cleanelastic');
+
+        parent::_construct();
+    }
+
+    /**
+     * @return \Elastica\Client
+     */
     public function getClient()
     {
         if (isset($this->_client)) {
@@ -27,10 +47,83 @@ class Clean_ElasticSearch_Model_Index extends Varien_Object
             return $this->_index;
         }
 
-        $elasticaIndex = $this->getClient()->getIndex(Mage::helper('cleanelastic')->getIndexName());
+        $elasticaIndex = $this->_initIndex();
 
         $this->_index = $elasticaIndex;
         return $this->_index;
+    }
+
+    public function _initIndex($new = false)
+    {
+        $store = Mage::app()->getStore();
+        $name = Mage::helper('cleanelastic')->getIndexName();
+        $index = $this->getClient()->getIndex($name);
+
+        // If index doesn't exist, create it with store settings
+        if (!$this->indexExists($name)) {
+            $index->create(array('settings' => $this->_helper->getStoreIndexSettings($store)));
+
+            // Send index mapping if not yet defined
+            foreach ($this->_helper->getStoreTypes($store) as $type) {
+                if ($index->getType($type)->exists()) {
+                    continue;
+                }
+                $mapping = new \Elastica\Type\Mapping();
+                $mapping->setType($index->getType($type));
+                if (!$this->isSourceEnabled($store)) {
+                    $mapping->disableSource();
+                }
+
+                // Hanle boost at query time
+                $properties = $this->getIndexer($type)->getStoreIndexProperties($store);
+                foreach ($properties as &$field) {
+                    if (isset($field['boost'])) {
+                        unset($field['boost']);
+                    }
+                }
+                unset($field);
+
+                $mapping->setAllField(array('analyzer' => 'std'));
+
+                $mapping->setProperties($properties);
+
+                Mage::dispatchEvent('clean_elasticsearch_mapping_send_before', array(
+                    'client' => $this,
+                    'store' => $store,
+                    'mapping' => $mapping,
+                    'type' => $type,
+                ));
+
+                $mapping->getType()->request(
+                    '_mapping',
+                    \Elastica\Request::PUT,
+                    $mapping->toArray(),
+                    array('update_all_types' => false)
+                );
+            }
+
+            // Set index analyzers for future search
+            //$index->setAnalyzers($this->_helper->getStoreAnalyzers($store));
+        }
+
+        return $index;
+    }
+
+    /**
+     * Checks if given index already exists
+     * Here because of a bug when calling exists() method directly on index object during index process
+     *
+     * @param mixed $index
+     * @return bool
+     */
+    public function indexExists($index)
+    {
+        return $this->getClient()->getStatus()->indexExists($index);
+    }
+
+    public function isSourceEnabled($store = null)
+    {
+        return true;
     }
 
     /**
@@ -52,7 +145,11 @@ class Clean_ElasticSearch_Model_Index extends Varien_Object
     public function deleteIndex()
     {
         try {
-            return $this->getIndex()->delete();
+            $name = Mage::helper('cleanelastic')->getIndexName();
+            if ($this->indexExists($name)) {
+                $this->getIndex()->delete();
+                unset($this->_index);
+            }
         } catch (Exception $e) {
             if (strpos($e->getMessage(), 'IndexMissingException') !== false) {
                 // Do nothing
@@ -60,5 +157,17 @@ class Clean_ElasticSearch_Model_Index extends Varien_Object
                 throw $e;
             }
         }
+
+        return $this;
+    }
+
+    /**
+     * @param $type string
+     * @return Clean_ElasticSearch_Model_IndexType_Abstract
+     */
+
+    public function getIndexer($type)
+    {
+        return Mage::getSingleton('cleanelastic/indexType_'.$type);
     }
 }
